libtcmu: move SCSI helpers out of libtcmu
These SCSI helpers depend on tcmu-runner to work so move them out
of libtcmu in this version. There is possibility to bring them or
some of them back in the future release.
