+++ libtcmu_common.h
@@ -70,12 +70,6 @@ enum {
 #define CFGFS_TARGET_MOD "/sys/module/target_core_user"
 #define CFGFS_MOD_PARAM CFGFS_TARGET_MOD"/parameters"
 
-/* Temporarily limit this to 32M */
-#define VPD_MAX_UNMAP_LBA_COUNT            65536
-#define VPD_MAX_UNMAP_BLOCK_DESC_COUNT     0x04
-/* Temporarily limit this is 0x1 */
-#define MAX_CAW_LENGTH                     0x01
-
 #define max(a, b) ({			\
 	__typeof__ (a) _a = (a);	\
 	__typeof__ (b) _b = (b);	\
@@ -99,8 +93,6 @@ enum {
 	(_a - (_a % _b)); })
 
 #define ARRAY_SIZE(arr) (sizeof(arr) / sizeof((arr)[0]))
-
-#define VPD_MAX_WRITE_SAME_LENGTH 0xFFFFFFFF
 
 typedef void (*cmd_done_t)(struct tcmu_device *, struct tcmulib_cmd *, int);
 
@@ -181,25 +173,10 @@ bool tcmu_zeroed_iovec(struct iovec *iovec, size_t iov_cnt);
 size_t tcmu_memcpy_into_iovec(struct iovec *iovec, size_t iov_cnt, void *src, size_t len);
 size_t tcmu_memcpy_from_iovec(void *dest, size_t len, struct iovec *iovec, size_t iov_cnt);
 size_t tcmu_iovec_length(struct iovec *iovec, size_t iov_cnt);
-bool char_to_hex(unsigned char *val, char c);
-
-/* Basic implementations of mandatory SCSI commands */
 int tcmu_set_sense_data(uint8_t *sense_buf, uint8_t key, uint16_t asc_ascq);
 void tcmu_set_sense_info(uint8_t *sense_buf, uint32_t info);
 void tcmu_set_sense_key_specific_info(uint8_t *sense_buf, uint16_t info);
 void __tcmu_set_sense_data(uint8_t *sense_buf, uint8_t key, uint16_t asc_ascq);
-int tcmu_emulate_inquiry(struct tcmu_device *dev, struct tgt_port *port, uint8_t *cdb, struct iovec *iovec, size_t iov_cnt);
-int tcmu_emulate_start_stop(struct tcmu_device *dev, uint8_t *cdb);
-int tcmu_emulate_test_unit_ready(uint8_t *cdb, struct iovec *iovec, size_t iov_cnt);
-int tcmu_emulate_read_capacity_10(uint64_t num_lbas, uint32_t block_size, uint8_t *cdb,
-				  struct iovec *iovec, size_t iov_cnt);
-int tcmu_emulate_read_capacity_16(uint64_t num_lbas, uint32_t block_size, uint8_t *cdb,
-				  struct iovec *iovec, size_t iov_cnt);
-int tcmu_emulate_mode_sense(struct tcmu_device *dev, uint8_t *cdb,
-			    struct iovec *iovec, size_t iov_cnt);
-int tcmu_emulate_mode_select(struct tcmu_device *dev, uint8_t *cdb,
-			     struct iovec *iovec, size_t iov_cnt);
-/* SCSI helpers */
 void tcmu_print_cdb_info(struct tcmu_device *dev, const struct tcmulib_cmd *cmd, const char *info);
 
 /*
