+++ tcmu-synthesizer.c
@@ -22,6 +22,7 @@
 #include "version.h"
 #include "libtcmu_config.h"
 #include "libtcmu_log.h"
+#include "scsi.h"
 
 typedef struct {
 	GIOChannel *gio;
